test: テストコードのコンパイルエラー・型不一致を全修正
- vital_service_test: model.Vital → model.VitalRecord, MedicalRecordID を *uint64 に修正, ptrUint64 重複定義を除去
- cage_service_test: price の型 float64 → int64 (model.Cage.Price は *int64)
- treatment_service_test: unitPrice/newUnitPrice の型 float64 → int64
- estimate_service_test: newAmount の型 float64 → int64
- treatment_plan_service_test: newPrice の型 float64 → int64
- daily_record_service_test: CreateStaffNoteRecordInput.Time に string → time.Time
- schema_drift_test: toSnakeCase を修正（連続大文字 ID → id 正常変換）, &model.Vital{} → &model.VitalRecord{}
- status-helpers.test: Notionパレットhex値に更新, getAccountingStatusColor のステータス値修正, getHospitalizationTypeColor の期待値修正

+++ backend/internal/service/cage_service_test.go
@@ -118,7 +118,7 @@ func TestCageService_List(t *testing.T) {
 }
 
 func TestCageService_GetByID(t *testing.T) {
-	price := 5000.0
+	price := int64(5000)
 	tests := []struct {
 		name     string
 		id       uint64
@@ -184,7 +184,7 @@ func TestCageService_GetByID(t *testing.T) {
 }
 
 func TestCageService_Create(t *testing.T) {
-	price := 3000.0
+	price := int64(3000)
 	tests := []struct {
 		name    string
 		cage    *model.Cage
@@ -249,7 +249,7 @@ func TestCageService_Create(t *testing.T) {
 }
 
 func TestCageService_Update(t *testing.T) {
-	price := 4500.0
+	price := int64(4500)
 	tests := []struct {
 		name    string
 		cage    *model.Cage
